Remove duplicate declaration of org_golang_x_tools.
With latest Bazel at HEAD, this results in an error, as rules_go already
imports the repository. If it is needed at a different version than
rules_go uses, the go_repository definition needs to be before the call
to go_repositories().

Also update the reference to io_bazel to the latest released version.